♻️ Move _get_quantity_for_epoch to top level
